Port board to Papilio Duo
Created a port of the HDMI-Light board to match the headers on a
Papilio Duo instead of a Papilio Pro. They both have the same FPGA, so
there should be no software changes required except changing the pin
assignment of the FPGA.

I added a footprint for an (optional) ZigBee module, since my plan is
to integrate HDMI-Light as a moodlight when there is no HDMI signal,
and have it be part of my ZigBee lighting network. If you don't
need/want that functionality, simply don't populate the module.

NOTE: In order to save cost, I kept the board deliberately under 100x50
mm, which is a limit for a few Chinese board houses. This means the
bottom row of headers on the Papilio Duo is not covered by the PCB, so
you can still access them as per usual. They might still be useful.